requirements.txt
alembic==1.3.1
aniso8601==8.0.0
argon2-cffi==19.2.0
cffi==1.13.2
Click==7.0
Flask==1.1.1
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
